% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_columns.R
\name{prepare_columns}
\alias{prepare_columns}
\title{Adds columns to data in preparation for feature generation using the epitopes package}
\usage{
prepare_columns(data)
}
\arguments{
\item{data}{a \emph{data.table}}
}
\value{
Data table of class \emph{windowed_epit_dt} containing the new columns.
}
\description{
Adds columns to data in preparation for feature generation using the epitopes package
}
\author{
Jodie Ashford (\email{ashfojsm@aston.ac.uk})
}
